Added logic to make realtor only able to delete homes. There may be a bug if I try and delete a home not created by a realtor

+++ src/home/home.controller.ts
@@ -77,7 +77,15 @@ export class HomeController {
   }
 
   @Delete(':id')
-  deleteHome(@Param('id', ParseIntPipe) id: number) {
+  async deleteHome(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserCategory,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException();
+    }
     return this.homeService.deleteHomeById(id);
   }
 }
